requirements.txt
openai
python_pptx
google_images_search
